hevelius data catalog renamed to hevelius task search
